feat(auth): rework auth hooks
