+++ packages/auth/src/index.ts
@@ -1,3 +1,6 @@
 export * from "./useAuthUser";
-export * from "./useAuthQuery";
-export * from "./useAuthMutation";
+export * from "./useAuthIdToken";
+export * from "./useAuthFetchSignInMethodsForEmail";
+export * from "./useAuthGetRedirectResult";
+
+export * from "./mutations";
